requirements.txt
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
scikit-learn>=1.3.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0
python-dotenv>=1.0.0
huggingface-hub>=0.19.0
